# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravelha API",
        "version": "0.0.1"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
        "/categories": {
            "get": {
                "tags": [
                    "/categories"
                ],
                "summary": "Display a listing of the resource",
                "description": "get all category on database and paginate then",
                "operationId": "App\\Http\\Controllers\\CategoryController::index",
                "parameters": [
                    {
                        "name": "only",
                        "in": "query",
                        "description": "filter results using field1;field2;field3...",
                        "style": "form",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "search results using key:value",
                        "style": "form",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "operators",
                        "in": "query",
                        "description": "set fileds operators using field1:operator1",
                        "style": "form",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "order results using field:direction",
                        "style": "form",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "with",
                        "in": "query",
                        "description": "get relations using relation1;relation2;relation3...",
                        "style": "form",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "define page",
                        "style": "form",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "limit per page",
                        "style": "form",
                        "schema": {
                            "type": "int"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of categories"
                    }
                }
            },
            "post": {
                "tags": [
                    "/categories"
                ],
                "summary": "Store a newly created resource in storage.",
                "description": "store a new category on database",
                "operationId": "App\\Http\\Controllers\\CategoryController::store",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "published_at": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "New category created"
                    }
                }
            }
        },
        "/categories/{category}": {
            "get": {
                "tags": [
                    "/categories"
                ],
                "summary": "Display the specified resource.",
                "description": "show category",
                "operationId": "App\\Http\\Controllers\\CategoryController::show",
                "parameters": [
                    {
                        "name": "category",
                        "in": "path",
                        "description": "Category id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "only",
                        "in": "query",
                        "description": "filter results using field1;field2;field3...",
                        "style": "form",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "with",
                        "in": "query",
                        "description": "get relations using relation1;relation2;relation3...",
                        "style": "form",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Show category"
                    }
                }
            },
            "put": {
                "tags": [
                    "/categories"
                ],
                "summary": "Update the specified resource in storage",
                "description": "update a category on database",
                "operationId": "App\\Http\\Controllers\\CategoryController::update",
                "parameters": [
                    {
                        "name": "category",
                        "in": "path",
                        "description": "Category id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "published_at": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Category updated"
                    }
                }
            },
            "delete": {
                "tags": [
                    "/categories"
                ],
                "summary": "Remove the specified resource from storage.",
                "description": "remove a category on database",
                "operationId": "App\\Http\\Controllers\\CategoryController::destroy",
                "parameters": [
                    {
                        "name": "category",
                        "in": "path",
                        "description": "Category id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Category deleted"
                    }
                }
            }
        },
        "/posts": {
            "get": {
                "tags": [
                    "/posts"
                ],
                "summary": "Display a listing of the resource",
                "description": "get all post on database and paginate then",
                "operationId": "App\\Http\\Controllers\\PostController::index",
                "parameters": [
                    {
                        "name": "only",
                        "in": "query",
                        "description": "filter results using field1;field2;field3...",
                        "style": "form",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "search results using key:value",
                        "style": "form",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "operators",
                        "in": "query",
                        "description": "set fileds operators using field1:operator1",
                        "style": "form",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "order results using field:direction",
                        "style": "form",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "with",
                        "in": "query",
                        "description": "get relations using relation1;relation2;relation3...",
                        "style": "form",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "define page",
                        "style": "form",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "limit per page",
                        "style": "form",
                        "schema": {
                            "type": "int"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of posts"
                    }
                }
            },
            "post": {
                "tags": [
                    "/posts"
                ],
                "summary": "Store a newly created resource in storage.",
                "description": "store a new post on database",
                "operationId": "App\\Http\\Controllers\\PostController::store",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "content": {
                                        "type": "string"
                                    },
                                    "published_at": {
                                        "type": "string"
                                    },
                                    "category_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "New post created"
                    }
                }
            }
        },
        "/posts/{post}": {
            "get": {
                "tags": [
                    "/posts"
                ],
                "summary": "Display the specified resource.",
                "description": "show post",
                "operationId": "App\\Http\\Controllers\\PostController::show",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "description": "Post id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "only",
                        "in": "query",
                        "description": "filter results using field1;field2;field3...",
                        "style": "form",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "with",
                        "in": "query",
                        "description": "get relations using relation1;relation2;relation3...",
                        "style": "form",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Show post"
                    }
                }
            },
            "put": {
                "tags": [
                    "/posts"
                ],
                "summary": "Update the specified resource in storage",
                "description": "update a post on database",
                "operationId": "App\\Http\\Controllers\\PostController::update",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "description": "Post id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "content": {
                                        "type": "string"
                                    },
                                    "published_at": {
                                        "type": "string"
                                    },
                                    "category_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Post updated"
                    }
                }
            },
            "delete": {
                "tags": [
                    "/posts"
                ],
                "summary": "Remove the specified resource from storage.",
                "description": "remove a post on database",
                "operationId": "App\\Http\\Controllers\\PostController::destroy",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "description": "Post id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Post deleted"
                    }
                }
            }
        },
        "/news/articles": {
            "get": {
                "tags": [
                    "/news/articles"
                ],
                "summary": "Display a listing of the resource",
                "description": "get all article on database and paginate then",
                "operationId": "Laravelha\\News\\Http\\Controllers\\ArticleController::index",
                "parameters": [
                    {
                        "name": "only",
                        "in": "query",
                        "description": "filter results using field1;field2;field3...",
                        "style": "form",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "search results using key:value",
                        "style": "form",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "operators",
                        "in": "query",
                        "description": "set fileds operators using field1:operator1",
                        "style": "form",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "order results using field:direction",
                        "style": "form",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "with",
                        "in": "query",
                        "description": "get relations using relation1;relation2;relation3...",
                        "style": "form",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "define page",
                        "style": "form",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "limit per page",
                        "style": "form",
                        "schema": {
                            "type": "int"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of articles"
                    }
                }
            },
            "post": {
                "tags": [
                    "/news/articles"
                ],
                "summary": "Store a newly created resource in storage.",
                "description": "store a new article on database",
                "operationId": "Laravelha\\News\\Http\\Controllers\\ArticleController::store",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "content": {
                                        "type": "string"
                                    },
                                    "published_at": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "New article created"
                    }
                }
            }
        },
        "/news/articles/{article}": {
            "get": {
                "tags": [
                    "/news/articles"
                ],
                "summary": "Display the specified resource.",
                "description": "show article",
                "operationId": "Laravelha\\News\\Http\\Controllers\\ArticleController::show",
                "parameters": [
                    {
                        "name": "article",
                        "in": "path",
                        "description": "Article id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "only",
                        "in": "query",
                        "description": "filter results using field1;field2;field3...",
                        "style": "form",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "with",
                        "in": "query",
                        "description": "get relations using relation1;relation2;relation3...",
                        "style": "form",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Show article"
                    }
                }
            },
            "put": {
                "tags": [
                    "/news/articles"
                ],
                "summary": "Update the specified resource in storage",
                "description": "update a article on database",
                "operationId": "Laravelha\\News\\Http\\Controllers\\ArticleController::update",
                "parameters": [
                    {
                        "name": "article",
                        "in": "path",
                        "description": "Article id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "content": {
                                        "type": "string"
                                    },
                                    "published_at": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Article updated"
                    }
                }
            },
            "delete": {
                "tags": [
                    "/news/articles"
                ],
                "summary": "Remove the specified resource from storage.",
                "description": "remove a article on database",
                "operationId": "Laravelha\\News\\Http\\Controllers\\ArticleController::destroy",
                "parameters": [
                    {
                        "name": "article",
                        "in": "path",
                        "description": "Article id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Article deleted"
                    }
                }
            }
        }
    }
}